requirements.txt
asgiref==3.8.1
asyncio==3.4.3
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
djangorestframework==3.15.2
idna==3.8
Pygments==2.18.0
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.1
urllib3==2.2.2
